Changed power from long to double.

+++ src/BacklightController/3x2.cpp
@@ -71,17 +71,37 @@ const size_t low_voltage_channels[LOW_VOLTAGE_COUNT] =
 // Units
 
 // Properties
-const long ir_backlight_power_max_default[IR_BACKLIGHT_COUNT] =
+// const long switching_frequency_max_default[CHANNEL_COUNT_MAX] =
+// {
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+//   switching_frequency_max,
+// };
+
+const double ir_backlight_power_max_default[IR_BACKLIGHT_COUNT] =
 {
   ir_backlight_power_max_element_default,
 };
 
-const long visible_backlight_power_max_default[VISIBLE_BACKLIGHT_COUNT] =
+const double visible_backlight_power_max_default[VISIBLE_BACKLIGHT_COUNT] =
 {
   visible_backlight_power_max_element_default,
 };
 
-const long high_voltage_power_max_default[HIGH_VOLTAGE_COUNT] =
+const double high_voltage_power_max_default[HIGH_VOLTAGE_COUNT] =
 {
   high_voltage_power_max_element_default,
   high_voltage_power_max_element_default,
@@ -89,7 +109,7 @@ const long high_voltage_power_max_default[HIGH_VOLTAGE_COUNT] =
   high_voltage_power_max_element_default,
 };
 
-const long low_voltage_power_max_default[LOW_VOLTAGE_COUNT] =
+const double low_voltage_power_max_default[LOW_VOLTAGE_COUNT] =
 {
   low_voltage_power_max_element_default,
   low_voltage_power_max_element_default,
